Refactor sidebar menu
