--- apps/web/components/auth/DesktopLayout/menuTabs.ts
@@ -1,1 +1,0 @@
-export const MenuTabs = ["Home", "Search", "Dashboard", "Profile"]
