+++ apps/web/components/dashboard/DashboardLayout/menuTabs.ts
@@ -1,0 +1,1 @@
+export type MenuTab = "home" | "search" | "dashboard" | "profile"
